Move password reset logic and code to user_authn.
