+++ common/djangoapps/student/message_types.py
@@ -7,13 +7,6 @@ from __future__ import absolute_import
 from openedx.core.djangoapps.ace_common.message import BaseMessageType
 
 
-class PasswordReset(BaseMessageType):
-    def __init__(self, *args, **kwargs):
-        super(PasswordReset, self).__init__(*args, **kwargs)
-
-        self.options['transactional'] = True
-
-
 class AccountRecovery(BaseMessageType):
     def __init__(self, *args, **kwargs):
         super(AccountRecovery, self).__init__(*args, **kwargs)
