+++ openedx/core/djangoapps/user_authn/templates/user_authn/edx_ace/passwordreset/email/from_name.txt
