+++ openedx/core/djangoapps/user_authn/urls_common.py
@@ -11,7 +11,7 @@ from __future__ import absolute_import
 from django.conf import settings
 from django.conf.urls import url
 
-from .views import auto_auth, login, logout, register
+from .views import auto_auth, login, logout, password_reset, register
 
 
 urlpatterns = [
@@ -19,7 +19,6 @@ urlpatterns = [
     url(r'^create_account$', register.RegistrationView.as_view(), name='create_account'),
     url(r'^user_api/v1/account/registration/$', register.RegistrationView.as_view(),
         name="user_api_registration"),
-
     # Login
     url(r'^login_post$', login.login_user, name='login_post'),
     url(r'^login_ajax$', login.login_user, name="login"),
@@ -31,6 +30,10 @@ urlpatterns = [
     url(r'^login_refresh$', login.login_refresh, name="login_refresh"),
 
     url(r'^logout$', logout.LogoutView.as_view(), name='logout'),
+
+    url(r'^v1/account/password_reset/$', password_reset.PasswordResetView.as_view(),
+        name="user_api_password_reset"),
+
 ]
