+++ openedx/core/djangoapps/user_authn/views/login_form.py
@@ -20,10 +20,10 @@ from openedx.core.djangoapps.site_configuration import helpers as configuration_
 from openedx.core.djangoapps.user_api.accounts.utils import is_secondary_email_feature_enabled
 from openedx.core.djangoapps.user_api.api import (
     get_login_session_form,
-    get_password_reset_form
 )
 from openedx.core.djangoapps.user_authn.cookies import are_logged_in_cookies_set
 from openedx.core.djangoapps.user_authn.views.registration_form import RegistrationFormFactory
+from openedx.core.djangoapps.user_authn.views.password_reset import get_password_reset_form
 from openedx.features.enterprise_support.api import enterprise_customer_for_request
 from openedx.features.enterprise_support.utils import (
     handle_enterprise_cookies_for_logistration,
